Places the parser-based sentence tokenizer besides the regex based
