+++ lib/natural/index.js
@@ -63,6 +63,7 @@ exports.WordPunctTokenizer = require('./tokenizers/regexp_tokenizer').WordPunctT
 exports.TreebankWordTokenizer = require('./tokenizers/treebank_word_tokenizer');
 exports.TokenizerJa = require('./tokenizers/tokenizer_ja');
 exports.SentenceTokenizer = require('./tokenizers/sentence_tokenizer');
+exports.SentenceTokenizerNew = require('./tokenizers/sentence_tokenizer_parser');
 exports.BayesClassifier = require('./classifiers/bayes_classifier');
 exports.LogisticRegressionClassifier = require('./classifiers/logistic_regression_classifier');
 exports.NounInflector = require('./inflectors/noun_inflector');
