+++ lib/natural/tokenizers/sentence_tokenizer.js
@@ -1,5 +1,5 @@
 /*
-Copyright (c) 2020, Hugo ter Doest
+Copyright (c) 2011, Chris Umbel
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -20,19 +20,32 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 */
 
-const parser = require('./parser_sentence_tokenizer')
-
 var Tokenizer = require('./tokenizer'),
     util = require('util');
+
+const DEBUG = false;
 
 var SentenceTokenizer = function() {
     Tokenizer.call(this);
 };
-
 util.inherits(SentenceTokenizer, Tokenizer);
 
 SentenceTokenizer.prototype.tokenize = function(text) {
-  return(parser.parse(text))
+    // break string up in to sentences based on punctation and quotation marks
+    var tokens = text.match(/(?<=\s+|^)[\"\'\‘\“\'\"\[\(\{\⟨](.*?[.?!])(\s[.?!])*[\"\'\’\”\'\"\]\)\}\⟩](?=\s+|$)|(?<=\s+|^)\S(.*?[.?!])(\s[.?!])*(?=\s+|$)/g);
+
+    DEBUG && console.log("SentenceTokenizer.tokenize: " + tokens);
+
+    if (!tokens) {
+        return [text];
+    }
+
+    // remove unecessary white space
+    tokens = tokens.map(Function.prototype.call, String.prototype.trim);
+
+    DEBUG && console.log("SentenceTokenizer.tokenize: tokens after removing whitespace " + tokens);
+
+    return this.trim(tokens);
 };
 
 module.exports = SentenceTokenizer;
